Did cleanups. Renamed folder: bl_material -> material
